feat: Change the client.Manager filter to be strict require token (#80)
When not providing a Bearer token the default implementation of Manager
will return a 401 Unauthorized error. The token contents are currently
not validated. With this change the Manager filter was removed from
AppBuilder as default filter to avoid other conflicts and needs to be
initiated manually in specific Webservices with the following snipet:

// inside Webservice Setup method
ws := &restful.WebService{}
if manager := client.ManagerCtx(ctx); manager != nil {
	ws = ws.Filter(manager.Filter(ctx))
}