Add SecretType to allow secret selecting with type
We used to only select secret with option ExcludedSecretTypes which
can help exclude secret which type in the ExcludedSecretTypes list.

Now I add SecretTypes means we only want secret which type in the
SecretTypes list.

It this option provided we will use it if not provided means there
is no limit for secret type when selecting

Signed-off-by: yuzhipeng <yuzp1996@qq.com>